HW3-change (not all) after pull comments
