--- src/Components/LoginPage/LoginPageComponents/SignButton.jsx
@@ -1,9 +1,0 @@
-import React from "react";
-
-export const SignButton = (props) => {
-    if (props.text) return (
-        <button type="submit" className="sign-button sign-button_style">{props.text}</button>
-    )
-
-    return ''
-}
