+++ src/components/Button.js
@@ -1,0 +1,6 @@
+import React from 'react';
+import {StyledButton} from "./ByttonStyles";
+
+export const Button = (props) => {
+    return <StyledButton {...props}/>
+};
