+++ src/components/Flex.js
@@ -1,0 +1,6 @@
+import React from 'react';
+import {StyledFlex} from "./FlexStyles";
+
+export const Flex = (props) => {
+    return <StyledFlex {...props}/>
+};
