+++ src/constants/dictionary.js
@@ -1,0 +1,8 @@
+export const restorePasswordTitle = `Restore Password`
+export const restorePasswordButtonText = 'Send Reset Link'
+
+export const SignInTitle = `Sign In`
+export const SignInButtonText = 'Sign In'
+
+export const SignUpTitle = `Sign Up`
+export const SignUpButtonText = 'Sign Up'
