+++ src/constants/doctorsFake.js
@@ -1,0 +1,20 @@
+export const doctorsFake = [
+    {
+        "avatar": "http://placeimg.com/640/480/abstract",
+        "firstName": "Bridgette",
+        "lastName": "Wisoky",
+        "profession": "Therapist",
+        "description": "Mollitia ex qui qui deleniti.\nEt accusamus doloribus laboriosam vitae.\nMaxime placeat repudiandae.",
+        "time": 1637840844,
+        "id": "4"
+    },
+    {
+        "avatar": "http://placeimg.com/640/480/abstract",
+        "firstName": "Will",
+        "lastName": "Hilll",
+        "profession": "Therapist",
+        "description": "Nulla debitis quia dignissimos veritatis facere. Minus iure aut aut facere incidunt quod minus. Excepturi est quo mollitia consequatur sequi corrupti magnam incidunt nesciunt. Et nesciunt est sed laudantium sed enim. Corrupti vel ullam qui saepe sit perspiciatis commodi ipsum. Est culpa nesciunt occaecati at id ut incidunt.\n \rAmet recusandae in eveniet quia in quas aut atque quia. Necessitatibus voluptas eum et pariatur quas nihil quod et voluptas. Qui voluptas nulla voluptates quos at.\n \rEt eius deserunt similique sit est iure. Et rerum modi quia consequatur voluptatem aspernatur. Necessitatibus adipisci et facilis molestias error sapiente minima. Tempora et aliquam id libero.",
+        "time": 1635840784,
+        "id": "5"
+    },
+]
