+++ src/pages/DoctorPage/Container/PatientCard/patientCard.css
