+++ src/pages/DoctorPage/Container/PatientCard/patientCard.less
