+++ src/pages/LoginPage/LoginPageComponents/Forms/RestorePasswordForm.jsx
@@ -4,12 +4,12 @@ import * as Yup from 'yup';
 import {SignFormInput} from "../SignFormInput";
 import {Link, useHistory} from "react-router-dom";
 import {RESTORE_PASSWORD_SENT_PATH, SIGN_IN_PATH} from "../../path";
+import {restorePasswordButtonText, restorePasswordTitle} from "../../../../constants/dictionary";
+import {restorePasswordValidationSchema} from "./validations";
 
 
 
 export const RestorePasswordForm = (props) => {
-    const signBarTitle = `Restore Password`
-    const signBarButtonText = 'Send Reset Link'
 
     const history = useHistory();
 
@@ -23,14 +23,9 @@ export const RestorePasswordForm = (props) => {
                 initialValues={{
                     email: '',
                 }}
-                validationSchema={Yup.object({
-                    email: Yup.string()
-                        .email('Invalid email address')
-                        .required('Required'),
-                })}
+                validationSchema={restorePasswordValidationSchema}
                 onSubmit={(values, {setSubmitting}) => {
                     setTimeout(() => {
-                        // alert(JSON.stringify(values, null, 2));
                         setSubmitting(false);
                         handleClick()
                     }, 400);
@@ -40,7 +35,7 @@ export const RestorePasswordForm = (props) => {
                 <Form className="sign-bar__form">
                     <div className="sign-bar__title">
                         <Link to={SIGN_IN_PATH} className="sign-bar__backArrow"> </Link>
-                        <div>{signBarTitle}</div>
+                        <div>{restorePasswordTitle}</div>
                     </div>
 
                     <p className={'sign-bar__text'}>Please use your email address, and we'll send you the link to reset
@@ -52,7 +47,7 @@ export const RestorePasswordForm = (props) => {
                         placeholder="Email"
                     />
 
-                    <button type="submit" className="sign-button sign-button_style">{signBarButtonText}</button>
+                    <button type="submit" className="sign-button sign-button_style">{restorePasswordButtonText}</button>
 
                 </Form>
             </Formik>
