+++ src/pages/LoginPage/LoginPageComponents/Forms/RestorePasswordSentForm.jsx
@@ -4,13 +4,12 @@ import {SIGN_IN_PATH} from "../../path";
 
 
 export const RestorePasswordSentForm = () => {
-    const signBarTitle = `Restore Password`
 
     return (
         <div className="sign-bar__form">
             <div className="sign-bar__title">
                 <Link to={SIGN_IN_PATH} className="sign-bar__backArrow"> </Link>
-                <div>{signBarTitle}</div>
+                <div>Restore Password</div>
             </div>
 
             <p className={'sign-bar__text'}>
