+++ src/pages/LoginPage/LoginPageComponents/Forms/SignInForm.jsx
@@ -3,14 +3,14 @@ import {Formik, Form} from 'formik';
 import * as Yup from 'yup';
 import {SignFormInput} from "../SignFormInput";
 import {Link} from "react-router-dom";
-import {RESTORE_PASSWORD_PATH, SIGN_IN_PATH} from "../../path";
-import { useHistory } from "react-router-dom";
+import {RESTORE_PASSWORD_PATH} from "../../path";
+import {useHistory} from "react-router-dom";
+import {SignInButtonText, SignInTitle} from "../../../../constants/dictionary";
+import {signInValidationSchema} from "./validations";
 
 
 
 export const SignInForm = (props) => {
-    const signBarTitle = `Sign In`
-    const signBarButtonText = 'Sign In'
 
     const history = useHistory();
 
@@ -25,20 +25,11 @@ export const SignInForm = (props) => {
                     email: '',
                     password: '',
                 }}
-                validationSchema={Yup.object({
-                    email: Yup.string()
-                        .email('Invalid email address')
-                        .required('Required'),
-                    password: Yup.string()
-                        .required('Required')
-                        .min(6, 'Must be 6 characters or more'),
-                })}
+                validationSchema={signInValidationSchema}
                 onSubmit={(values, {setSubmitting}) => {
                     setTimeout(() => {
-                        // alert(JSON.stringify(values, null, 2));
                         setSubmitting(false);
                         handleClick()
-                        // window.location.href = '/';
                     }, 400);
 
                 }}
@@ -46,7 +37,7 @@ export const SignInForm = (props) => {
 
                 <Form className="sign-bar__form">
                     <div>
-                        <div className="sign-bar__title">{signBarTitle}</div>
+                        <div className="sign-bar__title">{SignInTitle}</div>
                     </div>
 
                     <SignFormInput
@@ -63,7 +54,7 @@ export const SignInForm = (props) => {
                         toggleShowPassword={props.toggleShowPassword}
                     />
 
-                    <button type="submit" className="sign-button sign-button_style">{signBarButtonText}</button>
+                    <button type="submit" className="sign-button sign-button_style">{SignInButtonText}</button>
 
                     <div className={'sign-bar__forgotPassword'}>
                         <Link to={RESTORE_PASSWORD_PATH} className="sign-bar__forgotPassword">Forgot
