+++ src/pages/PatientPage/Container/DoctorCard/doctorCard.less
